Improve search path for _remote_common.sh
The remote_power script might be a symlink to the actual script. Before
this commit, it then failed to find _remote_common.sh. For example, a
local installation might look as follows:

	/usr/local/bin
	|-- remote_power -> /path/to/remote_power.git/remote_power
	|-- remote_power.cfg

Note that remote_power.cfg not being a symlink is a deliberate choice.
Being machine specific, it is not supposed to reside in the git repo.

Signed-off-by: Philip Oberfichtner <pro@denx.de>